change dialog behavior to be like a normal fragment
This trick uses for applying material theme for dialog fragments because it doesn't apply on dialog-fragments by default.
But here we're using it to change dialog behavior to be like a normal fragment.
Inspired from: https://dev.to/bhullnatik/how-to-use-material-dialogs-with-dialogfragment-28i1
It's important we should use `getView` in onViewCreated not view as argument