Fix NumberPicker view lagginess using android:layerType="hardware"
We have the same issue as https://stackoverflow.com/a/22554547 and
need the exactly same solution also apparently.

But the solution feels bad as doing so may expose GPU bugs
or maybe not? Perhaps not as there are instances of doing in
https://github.com/material-components/material-components-android/search?q=LAYER_TYPE_HARDWARE
but feel free to give me your feedbacks.

If so however, perhaps we need to redesign converter/gotodate's
design again to avoid android.widget.NumberPicker!

It feels weird that native system view like this would have such issues.