Fix back handling and onBackPressedDispatcher use
Basically the previous way of handling back button was one time,
it was enabled by default and was handling the event it wanted
once if made sense and otherwise it was disabling itself making
the callback no longer able to handle back pressing events
anymore.

This handles it by enable/disabling the callback itself each time
drawer or search view are in the position of needing to handling
back button themselves.