Add Jdn value class and move some of the related logics to it
We are currently passing Long values of Jdn and having scattered logic
about jdn value, this forms a no cost abstraction around the long value,
making the meaning of the passed values more clear and moves some of the
logic we had as top level functions in .utils package to it.

Value class has become stable in Kotlin 1.5 which is published just in some
days and the feature was introduced to me and suggested to use for Jdn values
by @beigirad and the upgrade to Kotlin 1.5 was done by @hrbdev in the same day
of the release.

Little by little the project should use this over passing Long values around yet
this removes these functions and turn them into a method to this new class,

* calculateWeekOfYear -> Jdn().getWeekOfYear()
* getDateFromJdnOfCalendar -> Jdn().toCalendar()
* getDateOfCalendar -> Jdn.fromDate()
* getTodayJdn -> Jdn.today
* getTodayOfCalendar -> Jdn.today.toCalendar()