Reuse arrow down of androidx.preferences in ArrowView
Preferences uses that and we are shipping it anyway so let's reuse
it in our own code also.

Its size looked a bit better also and on androidx.preferences removal
of some day we can revive our own arrow and make it based on this
but that won't happen this soon.