Schedule alarms both on WorkManager and AlarmManager
AlarmManager one work correctly on some devices and
not in some else, let's schedule alarms on both to make
we get called at least with one them.

Also implement a logic to make sure not the alarm is called
twice, the activity had the logic in itself and now that isn't needed.