Move all enabled events aggregation logic to SearchEventsAdapter
Now that we have more control over search adapter creation
let's do most of its logic off main-thread and remove the
scattered logic around aggregation of all enabled logics
into to the SearchEventsAdapter class itself.

This removes the global variable allEnabledEvents which
was there only for this purpose, paves the way for organizing
global variable of the project.