Revert "Implement an Android 12 inspired new UI"
This reverts commit 21f35046299c3131277bf7cd21d225b847b1bf1c.

Makes the UI laggish on some devices, let's postpone the change
to after release of Android 12 and limit it to higher Android
versions but maybe make it opt-in in newer Android versions.

Also UX concerns as this won't match with our current drawer menu
models, the drawer didn't open on sliding the screen from siding of
the device which should be addressed while the next try for
applying this.