astronomy: Separate minutesOffset updates from astronomyState
Both minutesOffset and astronomyState keep some sort of time state, astronomyState however
is meant to be used in animation thus is the visible one and the other is to keep final
animation value so we should keep the two in sync manually with a slight difference.

The separation has the benefit of not making reset button visible on initial animation
of the screen entrance and makes one day button to exactly jump 24h regardless of current
animation of the screen.