Remove dummy hidden RecyclerView of sliders
It is a bit unstable right now but works. It replaces EmptySlider which was
RecyclerView based and kinda hacky with a simple view that handles
touches and gestures itself.

It also can scroll in both axes (disabled by default) which was needed for
globe dialog but it isn't attached yet to y axis of texture sampler as I
didn't figure it out yet.

Needs further polishments but is good as is I can say.