Add MaterialFadeThrough to LicensesScreen
Apparently fragment also have a mechanism to set animation without
navigation library intevention and only that mechanism allows for
com.google.android.material.transition.Material* transition to be used.

We should investigate how we have best of it from them but in the
meanwhile let's add it to LicensesScreen as a corner place of the app
to not forget about the configuration.