Remove alarm manager scheduling and rely only on work manager
Another relics of the past, work manager itself is able to schedule using
alarm manager, let's just rely on that.

It initially was done because of older Android versions in initial versions of
work manager which shouldn't be the case anymore.

This specially is motivated by glance use of work manager as if that's what
glance uses and we have to adopt it anyway also.

Some may consider alarm manager more reliable than work manager which can
be possible but that's really the work manager and the device issue and
there were reports of users saying use of alarm manager will result in
battery usage warning and what not.