Remove multidex related bits
As Android 5 that's not a real concern anymore, though we probably
won't exceed 65k methods limits anytime soon but even so it's
that concern also so let's remove the legacy bits.

Also remove commented desugaring support, we didn't need it till
now and we can enable right when needed, TimeUnit to Duration change
feel superfluous and something we can go without.