week: A new screen
Currently it can be seen when week number feature is enabled and the
week numbers are clicked, or from the added icon in the schedule screen.

The code currently isn't the best and even the screen's name isn't decided yet.

Maybe it should have an entry on the calendar page's menu, maybe not.