Avoid use externalCacheDir for font and alarm
Cache dir is supposed for temporarily while we want to
store the font and the alarm persistently.

context.filesDir now is used for font file which is private to the app.

context.getExternalFilesDir now is used for storing the alarms
which system can access also.

Thanks to Amir for noticing this early, it could be disastrous if it
was exposed to the users.